Completely rebuilt config and web server
