--- package.json
@@ -1,8 +1,0 @@
-{
-  "name": "application-name",
-  "version": "0.0.1",
-  "dependencies": {
-    "redux-form": "latest",
-    "react": "latest"
-  }
-}
