+++ src/main/frontend/src/Application.less
@@ -1,0 +1,5 @@
+div.app-hello-message {
+  font-family: "Trebuchet MS", Verdana, sans-serif;
+  font-size: 20pt;
+  margin: 30pt;
+}
