+++ src/main/resources/static/index.html
@@ -1,17 +1,11 @@
 <!DOCTYPE html>
-
+<html>
 <head>
-    <meta charset="UTF-8" />
-    <title>Hello React!</title>
-    <script src="build/react.js"></script>
-    <script src="build/react-dom.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/babel-core/5.8.23/browser.min.js"></script>
+    <meta charset="utf-8" />
+    <title>Spring Boot + Maven Frontend Plugin + Node + Npm + Webpack + Babel + React + JQuery</title>
 </head>
 <body>
 <div id="content"></div>
-<script type="text/babel" src="app.js">
-</script>
-
-
+<script type="text/javascript" src="/bundle.js"></script>
 </body>
 </html>
