+++ src/test/java/digitalcitizen/ApplicationTests.java
@@ -1,0 +1,18 @@
+package digitalcitizen;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.test.context.web.WebAppConfiguration;
+import org.springframework.boot.test.SpringApplicationConfiguration;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+@RunWith(SpringJUnit4ClassRunner.class)
+@SpringApplicationConfiguration(classes = Application.class)
+@WebAppConfiguration
+public class ApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
